pyproject.toml
[project]
name = "database-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "langfuse>=3.11.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
